requirements.txt
python-dotenv
zeep
pydantic
requests
urllib3